Builds a simple sentinel reset tool
This package builds a simple tool to watch for pod deletions in a
redis-ha chart release and send a reset command to each sentinel. Per
the redis sentinel docs, this is done with 30 seconds between each
reset call. No reset is sent when deletion of the master pod (which also
has a sentinel on it) as this is typically going to be during master
failover which we do not want to trigger a reset during.